requirements.txt
requests~=2.26.0
logger
ConfigParser~=5.0.2
aniso8601==8.0.0
python-dotenv
click==8.0.3
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
python-owasp-zap-v2.4
redis
rstr
pyyaml
python-dateutil
PyJWT==1.7.1
sqlalchemy_utils
psycopg2 
openapi-spec-validator
black
semgrep
prance
openapi-core
fpdf
marshmallow
gunicorn